Correctly handle redirect on browser refresh
A logged in user visiting a secure route might refresh the browser. This action
wipes out the application state stored in Redux. Adding redux-persist persists
the state and rehydrates automatically. However, the browser refresh happened on
a secure route which triggers an authentication check and a possible redirect to
sign in. The timing of this check to inspect the store and the rehydration
cannot be relied upon to run in the correct order.

This commit issues an api call (for verification of authentication) by checking
the request object for the existance of a user. If the user exists, the
application state (i.e. Redux store) is updated and a redirect to sign in is
aborted.

If a logged in user opens a second browser tab, by default the Redux store will
be empty, yet the user will exist on the request object due to an existing
session.

This commit issues an api call (for verification of authentication) by checking
the request object for the existance of a user. If the user exists, the
application state (i.e. Redux store) is update and a redirect to the dashboard
is issued.