Reset wait time to 3. Remove check-db. Fix e2e no deck issue
Everything worked running on travis except for e2e tests that queried the db for
decks and cards.

The problem was that I had deck-import running first. Then unit and integration
tests cleaned out the database. When e2e tests run, they need the db to be
seeded with flashcard - and they were not.

So, I fixed the script in package.json to import decks prior to running e2e
tests.